Feat[world]: Add function pointer to world objects to resolve colissions
The idea is that once an object (the perpetrator) collides with another
object (the victim), the perpetrator runs its resolve_colission to
resolve the effects it applies to the victim. The victim also gets a
chance to 'defend' itself and runs its resolve_colission to apply
effects on the perpetrator.