Feat[engine]: Change executor for context
The context holds all of the environment of an app/game. Chose to remove
the executor and use the context method because:
- It allows having more than one app/game running on one program
- With the addition of events, it now needed another data structure; it
felt weird having another global variable for it.

The struct state_callback_t is mostly a workaround to be able to pass a
function pointer to a data pointer or in case there's a need to add
extra data to the callback later in the future.

The event_store uses the SDL_EventType enum; therefore it has 26 buckets
because it's the magic number that has the least amount of collisions
while using all the buckets.