Feat[engine/context]: Add event store and event dispatch/listener system.
The event store takes SDL_EventTypes to make a hashtable of dynamic
arrays where a callback and its target are stored.

For now, mouse events are the only ones that have a target resolver
function. Other events default to NULL (no target or global) since
there's no way to tell what the actual target is. Maybe someday in the
future, there can be target resolution for keyboard events or some other
ones.